requirements_local.txt
apache-airflow
apache-airflow[celery]
boto3
botocore
Flask
Flask-Admin
Flask-AppBuilder
Flask-Babel
Flask-Caching
Flask-Login
Flask-OpenID
Flask-SQLAlchemy
flask-swagger
Flask-WTF
funcsigs
future
gitdb2
GitPython
gunicorn
idna
numpy
pandas
pendulum
psutil
psycopg2-binary
pyspark
PyYAML
requests
s3transfer
setproctitle
six
SQLAlchemy
text-unidecode
thrift
tzlocal
unidecode
unicodecsv
urllib3
Werkzeug
WTForms
setuptools
pip3